# web/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_ingest_file_api_ingest_file_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_ingest_file_api_ingest_file_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InputType": {
        "enum": [
          "url",
          "pdf",
          "image",
          "text"
        ],
        "title": "InputType",
        "type": "string"
      },
      "JobDetailOut": {
        "description": "Extended job output that also embeds the list of draft recipe summaries.",
        "properties": {
          "artifacts": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Artifacts",
            "type": "object"
          },
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "cost_usd": {
            "default": 0.0,
            "title": "Cost Usd",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "drafts": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RecipeSummary"
            },
            "title": "Drafts",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "extraction_meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Meta"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_type": {
            "$ref": "#/components/schemas/InputType"
          },
          "produced_recipe_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Produced Recipe Ids",
            "type": "array"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "text_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Preview"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "id",
          "input_type",
          "status",
          "created_at",
          "attempts"
        ],
        "title": "JobDetailOut",
        "type": "object"
      },
      "JobOut": {
        "description": "Serialisable representation of a Job row.\n\n``source_fingerprint`` is intentionally EXCLUDED (internal implementation\ndetail \u2014 do not expose to clients).\n\n``url``, ``filename``, and ``text_preview`` are derived from *payload*\nat validation time (never the full raw payload).\n\n``artifacts`` maps store keys to /api/files/{ref} URL paths.\n``cost_usd`` is exposed as ``float`` for JSON friendliness.",
        "properties": {
          "artifacts": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Artifacts",
            "type": "object"
          },
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "cost_usd": {
            "default": 0.0,
            "title": "Cost Usd",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "extraction_meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Meta"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_type": {
            "$ref": "#/components/schemas/InputType"
          },
          "produced_recipe_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Produced Recipe Ids",
            "type": "array"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "text_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Preview"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "id",
          "input_type",
          "status",
          "created_at",
          "attempts"
        ],
        "title": "JobOut",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "needs_review",
          "done",
          "failed",
          "not_a_recipe"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "RecipeSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dish_types": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Dish Types",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Ref"
          },
          "is_verified": {
            "default": false,
            "title": "Is Verified",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "total_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Min"
          }
        },
        "required": [
          "id",
          "title",
          "created_at"
        ],
        "title": "RecipeSummary",
        "type": "object"
      },
...
      "SubmitTextIn": {
        "properties": {
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "SubmitTextIn",
        "type": "object"
      },
      "SubmitUrlIn": {
        "properties": {
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "SubmitUrlIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Recipe Normalizer",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ingest/file": {
      "post": {
        "operationId": "ingest_file_api_ingest_file_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_file_api_ingest_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest File",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/ingest/text": {
      "post": {
        "operationId": "ingest_text_api_ingest_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitTextIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Text",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/ingest/url": {
      "post": {
        "operationId": "ingest_url_api_ingest_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitUrlIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Url",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/jobs": {
      "get": {
        "operationId": "list_jobs_api_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobOut"
                  },
                  "title": "Response List Jobs Api Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/jobs/accept-high-confidence": {
      "post": {
        "operationId": "accept_high_confidence_api_jobs_accept_high_confidence_post",
        "parameters": [
          {
            "in": "query",
            "name": "threshold",
            "required": false,
            "schema": {
              "default": 0.9,
              "maximum": 1.0,
              "minimum": 0.0,
              "title": "Threshold",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Accept High Confidence Api Jobs Accept High Confidence Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept High Confidence",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/jobs/{job_id}/recipes/{recipe_id}/accept": {
      "post": {
        "operationId": "accept_draft_api_jobs__job_id__recipes__recipe_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "recipe_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recipe Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Draft",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/jobs/{job_id}/recipes/{recipe_id}/reject": {
      "post": {
        "operationId": "reject_draft_api_jobs__job_id__recipes__recipe_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "recipe_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recipe Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Draft",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/jobs/{job_id}/retry": {
      "post": {
        "operationId": "retry_job_api_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Job",
        "tags": [
          "ingestion"
        ]
      }
    },
...
  }
}